+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'level1'
+rootProject.name = 'level2'
