+++ src/main/java/com/innovation/level2/controller/BoardController.java
@@ -1,12 +1,11 @@
-package com.innovation.level1.controller;
+package com.innovation.level2.controller;
 
-import com.innovation.level1.dto.BoardDeleteDto;
-import com.innovation.level1.dto.BoardPatchDto;
-import com.innovation.level1.dto.BoardPostDto;
-import com.innovation.level1.dto.BoardResponseDto;
-import com.innovation.level1.entity.Board;
-import com.innovation.level1.mapper.BoardMapper;
-import com.innovation.level1.service.BoardService;
+import com.innovation.level2.dto.BoardDeleteDto;
+import com.innovation.level2.dto.BoardPatchDto;
+import com.innovation.level2.dto.BoardPostDto;
+import com.innovation.level2.dto.BoardResponseDto;
+import com.innovation.level2.entity.Board;
+import com.innovation.level2.service.BoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
