+++ src/main/java/com/innovation/level2/dto/BoardResponseDto.java
@@ -1,13 +1,14 @@
-package com.innovation.level1.dto;
+package com.innovation.level2.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
-@Getter @Setter
-public class BoardPatchDto {
+@Getter
+@Setter
+public class BoardResponseDto {
     private long boardId;
     private String title;
     private String username;
-    private String password;
     private String content;
+    private String createdAt;
 }
